added new card task component
